+++ ChangeLog.md
@@ -31,6 +31,7 @@ and simplify code. See the Venice *JMH* benchmark *ReflectionBenchmark.java*
 
 ### Changed
 
+- function `io/download` to accept a progress function as option
 - optional OpenPDF dependency to version 1.3.16